Basic Movement
I have created very basic movement for the player object. I am not an expert on this and do not understand half of what I did so if any of this is messing up anything else, we can find a different way to do it and do not need to push this to main.