Changes to Movement
Previously, the player's force was repeatedly modified, meaning that actual force-based stuff was never updating (like gravity). I've changed it so it has a similar behavior and you can turn and stop instantly without having to worry about friction.
Along with this, I've modified it so the character actually starts and stops walking and is animation properly.
moveInput.y is also no longer zero, because we will have to read this if we include stuff like ladders.